@@ -1,16 +1,16 @@
 {
   "name": "@knaydenov/material-grid",
-  "version": "1.0.18",
+  "version": "2.0.0",
   "peerDependencies": {
-    "@angular/cdk": "^6.4.0",
-    "@angular/common": "^6.0.0-rc.0 || ^6.0.0",
-    "@angular/core": "^6.0.0-rc.0 || ^6.0.0",
-    "@angular/material": "^6.4.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/material": "^8.0.0",
     "@ngx-translate/core": "^10.0.2"
   },
   "devDependencies": {
-    "@angular/cdk": "^6.4.0",
-    "@angular/material": "^6.4.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/material": "^8.0.0",
     "@ngx-translate/core": "^10.0.2",
     "scss-bundle": "^2.3.2"
   }
